fix: cannot resize partition XXX Not enough space
If you ever come into a situation where you cannot upgrade anymore due to a resize error [2]
you need to disable avb-verity and enable it again. The reason is that as soon as you have ever used 
the "adb remount" command a partition named "scratch" will be created [3]. Disabling verity and enabling it again
will remove that partition and so fix that issue.

BUT: this does not work on USER builds as avb-verity commands will be rejected.
so I found another solution which requires to build a factory image which creates a "super.img" [1]

1. set/change "PRODUCT_BUILD_SUPER_PARTITION := true" e.g. in your device.mk (and ensure its not set to "false" anywhere else)
2. If you have a device with PRODUCT_SHIPPING_API_LEVEL set to PIE or lower you can just set this in your BoardConfig:
`PRODUCT_RETROFIT_DYNAMIC_PARTITIONS := true`
This won't work anymore when your shipping level is higher so you need to mod the source like that:
`build/core/Makefile` -> search for the PHONY entry: "superimage_dist"
AFTER that entry you should find a line `endif # PRODUCT_RETROFIT_DYNAMIC_PARTITIONS != "true"`
-> outcomment it
BEFORE that PHONY entry -> some lines up there should be: "ifneq (true,$(PRODUCT_RETROFIT_DYNAMIC_PARTITIONS))" 
-> outcomment it, too
3. build the super.img with `mka / make superimage_dist`
4. flash it:
a) fastboot: boot your device in fastboot and flash it to the super partition
b) TWRP/SHRP: push the super.img to the device, select flash, select the super.img and flash it to "super" -> enjoy

[1](https://source.android.com/devices/tech/ota/dynamic_partitions/implement#factory-images)
[2] Example errors:

[liblp]Partition vendor will resize from 0 bytes to 217198592 bytes
[liblp]Not enough free space to expand partition: system
Failed to resize partition system to size 1719984128.

or:

Cannot resize partition product_b to size 2570489856. Not enough space?

[3](https://source.android.com/devices/tech/ota/dynamic_partitions/implement#adb-remount)